// front-end/src/modules/user/Projects/ProjectList/projects.types.ts
export interface IProjectListDetails {
  company_id: string;
  country: string;
  head_contract_sum: string;
  id: string;
  latitude: string;
  longitude: string;
  number_of_units: string;
  place_id: string;
  project_date: string;
  project_description: string;
  project_id: string;
  project_name: string;
  project_role: string;
  project_status: string;
  pta_eligibility: string;
  region: string;
  retention_type: string;
  rta_eligibility: string;
  site_address: string;
}